server/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Flask-JWT-Extended>=4.6.0
passlib>=1.7.4
openai>=1.51.0
anthropic>=0.39.0
resend>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
pypdf>=5.1.0
python-docx>=1.1.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0

autopep8>=2.0.0
